move org chart components to relative parent folder
